pyproject.toml
[project]
name = "noteshift"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Luke", email = "lukemaxwellshouse@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "python-slugify>=8.0.4",
    "rich>=14.3.3",
    "typer>=0.24.1",
]

[project.scripts]
noteshift = "noteshift.cli:main"

[tool.ruff]
line-length = 110
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]

[build-system]
requires = ["uv_build>=0.10.4,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.4",
]
